fix: Refactor Header buttons to make them consistent on Android and iOS.
